pyproject.toml
[tool.poetry]
name = "langchain-tutorial"
version = "0.1.0"
description = ""
authors = ["Akira Abe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "0.1.6"
boto3 = "1.34.39"
python-dotenv = "1.0.1"
langchain-community = "0.0.19"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
